refactor: split local-first store responsibilities
